removing shared function
